backend/requirements.txt
fastapi
uvicorn[standard]
numpy
librosa
soundfile
tensorflow          # or tensorflow-cpu
python-multipart
pydantic
pillow
matplotlib
yt-dlp
